// types/rules.ts
export enum RuleTypeEnum {
  DOMAIN = "DOMAIN",
  DOMAIN_SUFFIX = "DOMAIN-SUFFIX",
  DOMAIN_KEYWORD = "DOMAIN-KEYWORD",
  DOMAIN_REGEX = "DOMAIN-REGEX",
  GEOSITE = "GEOSITE",
  IP_CIDR = "IP-CIDR",
  IP_CIDR6 = "IP-CIDR6",
  IP_SUFFIX = "IP-SUFFIX",
  IP_ASN = "IP-ASN",
  GEOIP = "GEOIP",
  SRC_GEOIP = "SRC-GEOIP",
  SRC_IP_ASN = "SRC-IP-ASN",
  SRC_IP_CIDR = "SRC-IP-CIDR",
  SRC_IP_SUFFIX = "SRC-IP-SUFFIX",
  DST_PORT = "DST-PORT",
  SRC_PORT = "SRC-PORT",
  IN_PORT = "IN-PORT",
  IN_TYPE = "IN-TYPE",
  IN_USER = "IN-USER",
  IN_NAME = "IN-NAME",
  PROCESS_PATH = "PROCESS-PATH",
  PROCESS_PATH_REGEX = "PROCESS-PATH-REGEX",
  PROCESS_NAME = "PROCESS-NAME",
  PROCESS_NAME_REGEX = "PROCESS-NAME-REGEX",
  UID = "UID",
  NETWORK = "NETWORK",
  DSCP = "DSCP",
  RULE_SET = "RULE-SET",
  AND = "AND",
  OR = "OR",
  NOT = "NOT",
  SUB_RULE = "SUB-RULE",
  MATCH = "MATCH"
}

export enum RuleActionEnum {
  REJECT = "REJECT",
  REJECT_DROP = "REJECT-DROP",
  DIRECT = "DIRECT",
  NO_RESOLVE = "no-resolve",
}